requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pydantic
mysql-connector-python
fastapi
jinja2
email-validator
